requirements.txt
Django==4.2.7
django-ckeditor==6.7.0
django-filter==23.3
django-js-asset==2.1.0
django-parler==2.3
django-parler-rest==2.2
django-rosetta==0.9.9
djangorestframework==3.14.0
Markdown==3.5.1
Pillow==10.1.0
django-cors-headers==4.3.1
drf-yasg==1.21.7 
inflection==0.5.1 
packaging==23.2 
pyyaml==6.0.1 
uritemplate==4.1.1
black==23.11.0